Cleaning up dsl and encoders

+++ src/main/scala/com/gardenShare/gardenshare/PaymentFee/DeterminePaymentFee.scala
@@ -1,7 +1,7 @@
 package com.gardenShare.gardenshare
 
 import cats.effect.IO
-
+import com.gardenShare.gardenshare.Price.PriceOps
 
 abstract class DeterminePaymentFee[F[_]] {
   def determineFee(amount: Amount, sellerEmail: Email): F[Amount]
@@ -9,6 +9,6 @@ abstract class DeterminePaymentFee[F[_]] {
 
 object DeterminePaymentFee {
   implicit object DefaultIODeterminePaymentFee extends DeterminePaymentFee[IO] {
-    def determineFee(amount: Amount, sellerEmail: Email): IO[Amount] = IO.pure(Amount((amount.quantityOfCurrency * .05).toInt, amount.currencyType))
+    def determineFee(amount: Amount, sellerEmail: Email): IO[Amount] = IO.pure(Amount(amount.quantityOfCurrency * .05, amount.currencyType))
   }
 }
